# docs/swagger.yaml
definitions:
  handlers.AttendanceFreezeInput:
    properties:
      comment:
        type: string
      end_date:
        type: string
      start_date:
        type: string
    type: object
  handlers.AttendanceLessonInput:
    properties:
      curator_id:
        type: string
      date:
        type: string
      feedback:
        type: string
      feedback_date:
        type: string
      format:
        type: string
      lessons_status:
        type: string
    required:
    - lessons_status
    type: object
  handlers.AttendanceProlongationInput:
    properties:
      amount:
        type: number
      comment:
        type: string
      date:
        type: string
      payment_type:
        enum:
        - оплата
        - предоплата
        - доплата
        type: string
    required:
    - payment_type
    type: object
...
  handlers.CreateAttendanceRequest:
    properties:
      course_id:
        type: string
      freeze:
        $ref: '#/definitions/handlers.AttendanceFreezeInput'
      lesson:
        $ref: '#/definitions/handlers.AttendanceLessonInput'
      prolongation:
        $ref: '#/definitions/handlers.AttendanceProlongationInput'
      student_id:
        type: string
      type:
        enum:
        - урок
        - заморозка
        - пролонгация
        type: string
    required:
    - course_id
    - student_id
    - type
    type: object
...
  models.ApiError:
    properties:
      error:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /attendances:
    post:
      consumes:
      - application/json
      description: |-
        Добавляет новую запись: урок, заморозку или пролонгацию
        Допустимые значения:
        - type: урок, заморозка, пролонгация
        - lessons_status: пропущен, проведен, запланирован, отменен
        - payment_type: оплата, предоплата, доплата
      parameters:
      - description: Данные посещаемости
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateAttendanceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать запись посещаемости
      tags:
      - Attendance
  /attendances/{id}:
    put:
      consumes:
      - application/json
      description: |-
        Обновляет запись посещаемости (урок, заморозка или пролонгация)
        Допустимые значения:
        - type: урок, заморозка, пролонгация
        - lessons_status: пропущен, проведен, запланирован, отменен
        - payment_type: оплата, предоплата, доплата
      parameters:
      - description: ID записи посещаемости
        in: path
        name: id
        required: true
        type: string
      - description: Обновленные данные посещаемости
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateAttendanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Обновить запись посещаемости
      tags:
      - Attendance
  /attendances/{studentId}:
    get:
      consumes:
      - application/json
      description: Возвращает список всех посещений, заморозок и пролонгаций по студенту
      parameters:
      - description: UUID студента
        in: path
        name: studentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items: {}
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Получить посещаемость студента
      tags:
      - Attendance
...
  /settings/attendance/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет запись посещаемости по ID
      parameters:
      - description: ID записи посещаемости
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      summary: Удалить запись посещаемости
      tags:
      - Attendance
...
swagger: "2.0"